@@ -1,18 +1,21 @@
 [build-system]
 build-backend = "setuptools.build_meta"
-requires = ["setuptools>=61", "setuptools-scm>=8.0"]
+requires = ["setuptools>=75.8.0", "setuptools-scm>=8.1.0"]
 
 [project]
 authors = [
   {name = "Jevin Sweval", email = "jevinsweval@gmail.com"}
 ]
 classifiers = [
-  "Programming Language :: Python :: 3"
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.12"
 ]
 dependencies = [
-  "path",
-  "rich",
-  "pre-commit"
+  "path>=17.1.0",
+  "rich>=13.9.4",
+  "pre-commit>=2.9.2"
 ]
 description = "dune format-dune-file pre-commit hook for dune and dune-project files"
 dynamic = ["version"]
@@ -20,7 +23,7 @@ keywords = ["dune", "dune-project", "dune-file", "dune-format", "format-dune-fil
 license = {text = "BSD-2-Clause"}
 name = "pre-commit-dune-format-dune-file"
 readme = {file = 'README.md', content-type = 'text/markdown'}
-requires-python = ">=3.9"
+requires-python = ">=3.12"
 
 [project.scripts]
 pre-commit-dune-format-dune-file = "pre_commit_dune_format_dune_file.format:main"
